:sparkles: URL 파라미터를 통해 유저 판별 상태 관리 Context 생성하여 Props 내려주는 것 방지
